+++ src/main/java/com/metavirtual/bloom/myPage/adminPage/controller/AdminPageController.java
@@ -1,4 +1,4 @@
-package com.metavirtual.bloom.myPage.adminpage.controller;
+package com.metavirtual.bloom.myPage.adminPage.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
