+++ src/main/java/com/metavirtual/bloom/myPage/adminPage/dao/AdminPageMapper.java
@@ -1,0 +1,4 @@
+package com.metavirtual.bloom.myPage.adminPage.dao;
+
+public class AdminPageMapper {
+}
