+++ src/main/java/com/metavirtual/bloom/myPage/adminPage/service/AdminPageService.java
@@ -1,0 +1,5 @@
+package com.metavirtual.bloom.myPage.adminPage.service;
+
+public class AdminPageService {
+
+}
