+++ src/main/java/com/metavirtual/bloom/myPage/memberPage/dao/MemberPageMapper.java
@@ -1,0 +1,4 @@
+package com.metavirtual.bloom.myPage.memberPage.dao;
+
+public class MemberPageMapper {
+}
