+++ src/main/java/com/metavirtual/bloom/myPage/memberPage/service/MemberPageService.java
@@ -1,0 +1,7 @@
+package com.metavirtual.bloom.myPage.memberPage.service;
+
+import com.metavirtual.bloom.user.dto.MemberDTO;
+
+public class MemberPageService {
+//    public void modifyMemberInfo(MemberDTO member, UserInfo_DTO userInfo);
+}
