+++ src/main/java/com/metavirtual/bloom/myPage/therapistPage/dao/TherapistPageMapper.java
@@ -1,0 +1,4 @@
+package com.metavirtual.bloom.myPage.therapistPage.dao;
+
+public class TherapistPageMapper {
+}
