+++ src/main/java/com/metavirtual/bloom/myPage/therapistPage/service/TherapistPageService.java
@@ -1,0 +1,4 @@
+package com.metavirtual.bloom.myPage.therapistPage.service;
+
+public class TherapistPageService {
+}
