+++ src/main/resources/templates/mypage/member/profile2.html
@@ -37,11 +37,11 @@
             <td style="border-bottom: 2px solid #C09A81;">010-1111-1234</td>
             <td style="border-bottom: 2px solid #C09A81;">예약 확정</td>
         </tr>
-<!--        <tr th:each="therapist : ${ therapistList }">-->
-<!--            <td th:text="${ therapistName }"></td>-->
-<!--            <td th:text="${ reserveDate }"></td>-->
-<!--            <td th:text="${ therapistPhone }"></td>-->
-<!--            <td th:text="${ reserveStatus }" style="border-radius: 10px; background-color: #C09A81; color: white;"></td>-->
+<!--        <tr align="center">-->
+<!--            <td th:if="${} th:text="${ therapistName }"></td>-->
+<!--            <td th:if="${} th:text="${ booking.bookingDate }"></td>-->
+<!--            <td th:if="${} th:text="${ user. }"></td>-->
+<!--            <td th:if="${} th:text="${ booking.bookingStatus }" style="border-radius: 10px; background-color: #C09A81; color: white;"></td>-->
 <!--        </tr>-->
     </table>
     <br>
